ai-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.57.0
pinecone-client==5.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.0
